Handle Clickhouse queries with other statements being invalid
If the original query is

```
insert into cats (name) values (' + var + ');
```

And the user input is

```
foo'||version());
```

Then the query will be

```sql
insert into cats_table (id, petname) values (null, 'foo'||version());');
```

Which isn't a valid query. The second statement is `');`

Clickhouse seems to ignore this because it does not allow multiple
statements. It will still execute the first part so we need to detect
this.